ELK and Cloud machine Summary
The documentation written provides what is in this repository. 
Using Microsoft Azure to create a cloud server with a monitoring server and two webservers. This server also provides load balancer and a jumpbox as security to certain attacks. 
Playbooks that were also implemented to monitor the webservers using a ELK system.